requirements.txt
APScheduler==3.10.1
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
greenlet==2.0.2
idna==3.4
loguru==0.7.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
win32-setctime==1.1.0
